+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## [Unreleased]
 ### Added:
 - 'Export to library file' Context menu to Favorites panel
+- 'Import from library file' Context menu to Favorites panel